Add DeviceContactProviderDataSourceTest.kt
