+++ app/build.gradle.kts
@@ -126,11 +126,10 @@ dependencies {
     implementation(project(":feature:onecontactwidget"))
     implementation(project(":feature:contactconfig"))
     implementation(project(":feature:settings"))
-    implementation(project(":data:contacts:impl:default"))
     implementation(project(":data:phones:api"))
-    implementation(project(":data:phones:impl:device"))
     implementation(project(":data:pictures:impl:default"))
     implementation(project(":data2"))
+    implementation(project(":data3:devicecontact"))
     implementation(project(":core:ui"))
     implementation(project(":core:analytics"))
     implementation(project(":core:di"))
