+++ app/src/main/res/layout/fragment_widget_config.xml
@@ -28,6 +28,10 @@
             type="com.blaxsoftware.directcallwidget.viewmodel.WidgetConfigViewModel" />
     </data>
 
+    <View
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" /><!--
+
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
@@ -129,5 +133,5 @@
                 </com.google.android.material.textfield.TextInputLayout>
             </androidx.constraintlayout.widget.ConstraintLayout>
         </androidx.core.widget.NestedScrollView>
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
+    </androidx.coordinatorlayout.widget.CoordinatorLayout>-->
 </layout>
