--- data/phones/impl/device/.gitignore
@@ -1,1 +1,0 @@
-/build
