+++ data3/devicecontact/proguard-rules.pro
