+++ data3/devicecontact/src/main/AndroidManifest.xml
